Remove redundant Property-related codes
