+++ include/neug/storages/graph/schema.h
@@ -28,6 +28,7 @@
 #include "neug/execution/common/types/value.h"
 #include "neug/utils/bitset.h"
 #include "neug/utils/id_indexer.h"
+#include "neug/utils/property/property.h"
 #include "neug/utils/property/types.h"
 #include "neug/utils/result.h"
 #include "neug/utils/serialization/in_archive.h"
