+++ tests/storage/test_csr_batch_ops.cc
@@ -2,6 +2,7 @@
 #include <filesystem>
 
 #include "neug/config.h"
+#include "neug/execution/common/types/value.h"
 #include "neug/storages/checkpoint_manager.h"
 #include "neug/storages/csr/immutable_csr.h"
 #include "neug/storages/csr/mutable_csr.h"
@@ -51,7 +52,8 @@ class CsrBatchTest : public ::testing::Test {
         actual;
     auto view = this->csr->get_generic_view(0);
     auto ed_accessor = neug::EdgeDataAccessor(
-        neug::PropUtils<typename T::data_t>::prop_type(), nullptr);
+        neug::execution::ValueConverter<typename T::data_t>::type().id(),
+        nullptr);
     for (neug::vid_t src = 0; src < this->csr->size(); ++src) {
       auto es = view.get_edges(src);
       for (auto it = es.begin(); it != es.end(); ++it) {
